+++ lib/pages/home_tab.dart
@@ -1,0 +1,18 @@
+// home_tab.dart
+
+import 'package:flutter/material.dart';
+
+class HomeTab extends StatelessWidget {
+  const HomeTab({Key? key}) : super(key: key);
+
+  @override
+  Widget build(BuildContext context) {
+    // Placeholder content for the Home tab
+    return Center(
+      child: Text(
+        'Welcome to the Home Tab',
+        style: TextStyle(fontSize: 24),
+      ),
+    );
+  }
+}
