+++ lib/pages/phone_sign_in_screen.dart
@@ -1,0 +1,153 @@
+// phone_sign_in_screen.dart
+
+import 'package:cloud_firestore/cloud_firestore.dart';
+import 'package:firebase_auth/firebase_auth.dart';
+import 'package:flutter/material.dart';
+
+import 'main_screen.dart';
+
+class PhoneSignInScreen extends StatefulWidget {
+  @override
+  _PhoneSignInScreenState createState() => _PhoneSignInScreenState();
+}
+
+class _PhoneSignInScreenState extends State<PhoneSignInScreen> {
+  final FirebaseAuth _auth = FirebaseAuth.instance;
+  final FirebaseFirestore _firestore = FirebaseFirestore.instance;
+
+  TextEditingController _phoneNumberController = TextEditingController();
+  TextEditingController _smsCodeController = TextEditingController();
+
+  String? _verificationId;
+
+  @override
+  void dispose() {
+    _phoneNumberController.dispose();
+    _smsCodeController.dispose();
+    super.dispose();
+  }
+
+  void _sendCodeToPhoneNumber() async {
+    String phoneNumber = _phoneNumberController.text.trim();
+    if (phoneNumber.isEmpty) {
+      ScaffoldMessenger.of(context)
+          .showSnackBar(SnackBar(content: Text('Please enter phone number')));
+      return;
+    }
+    await _auth.verifyPhoneNumber(
+      phoneNumber: phoneNumber,
+      verificationCompleted: (PhoneAuthCredential credential) async {
+        // Auto-sign in on Android
+        await _auth.signInWithCredential(credential);
+        await _saveUserToFirestore();
+      },
+      verificationFailed: (FirebaseAuthException e) {
+        ScaffoldMessenger.of(context).showSnackBar(
+            SnackBar(content: Text('Verification failed: ${e.message}')));
+      },
+      codeSent: (String verificationId, int? resendToken) {
+        setState(() {
+          _verificationId = verificationId;
+        });
+        ScaffoldMessenger.of(context)
+            .showSnackBar(SnackBar(content: Text('Verification code sent')));
+      },
+      codeAutoRetrievalTimeout: (String verificationId) {
+        _verificationId = verificationId;
+      },
+    );
+  }
+
+  void _signInWithPhoneNumber() async {
+    String smsCode = _smsCodeController.text.trim();
+    if (_verificationId == null) {
+      ScaffoldMessenger.of(context).showSnackBar(
+          SnackBar(content: Text('Please request verification code first')));
+      return;
+    }
+    if (smsCode.isEmpty) {
+      ScaffoldMessenger.of(context).showSnackBar(
+          SnackBar(content: Text('Please enter verification code')));
+      return;
+    }
+    try {
+      PhoneAuthCredential credential = PhoneAuthProvider.credential(
+          verificationId: _verificationId!, smsCode: smsCode);
+      await _auth.signInWithCredential(credential);
+      await _saveUserToFirestore();
+    } catch (e) {
+      ScaffoldMessenger.of(context).showSnackBar(
+          SnackBar(content: Text('Failed to sign in: ${e.toString()}')));
+    }
+  }
+
+  Future<void> _saveUserToFirestore() async {
+    User? user = _auth.currentUser;
+    if (user != null) {
+      // Store user data in Firestore without overwriting existing fields
+      await _firestore.collection('users').doc(user.uid).set({
+        'phoneNumber': user.phoneNumber,
+        'lastSignInTime': FieldValue.serverTimestamp(),
+        'provider': 'phone',
+      }, SetOptions(merge: true));
+
+      // Navigate to the main screen after successful sign-in
+      Navigator.of(context).pushAndRemoveUntil(
+        MaterialPageRoute(builder: (context) => MainScreen()),
+        (Route<dynamic> route) => false,
+      );
+    }
+  }
+
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      appBar: AppBar(
+        title: Text('Phone Sign-In'),
+      ),
+      body: Center(
+        child: SingleChildScrollView(
+          child: Column(
+            mainAxisAlignment: MainAxisAlignment.center,
+            children: [
+              // Phone Number Input
+              Padding(
+                padding: const EdgeInsets.symmetric(horizontal: 24.0),
+                child: TextField(
+                  controller: _phoneNumberController,
+                  decoration: InputDecoration(
+                    labelText: 'Phone Number',
+                    hintText: '+1234567890',
+                  ),
+                  keyboardType: TextInputType.phone,
+                ),
+              ),
+              SizedBox(height: 10),
+              ElevatedButton(
+                onPressed: _sendCodeToPhoneNumber,
+                child: Text('Send Code'),
+              ),
+              SizedBox(height: 10),
+              // Verification Code Input
+              Padding(
+                padding: const EdgeInsets.symmetric(horizontal: 24.0),
+                child: TextField(
+                  controller: _smsCodeController,
+                  decoration: InputDecoration(
+                    labelText: 'Verification Code',
+                  ),
+                  keyboardType: TextInputType.number,
+                ),
+              ),
+              SizedBox(height: 10),
+              ElevatedButton(
+                onPressed: _signInWithPhoneNumber,
+                child: Text('Verify Code'),
+              ),
+            ],
+          ),
+        ),
+      ),
+    );
+  }
+}
