+++ lib/pages/profile_tab.dart
@@ -1,0 +1,178 @@
+// profile_tab.dart
+
+import 'package:cloud_firestore/cloud_firestore.dart';
+import 'package:firebase_auth/firebase_auth.dart';
+import 'package:flutter/material.dart';
+
+import 'first_screen.dart';
+
+class ProfileTab extends StatefulWidget {
+  const ProfileTab({Key? key}) : super(key: key);
+
+  @override
+  _ProfileTabState createState() => _ProfileTabState();
+}
+
+class _ProfileTabState extends State<ProfileTab> {
+  final FirebaseAuth _auth = FirebaseAuth.instance;
+  final FirebaseFirestore _firestore = FirebaseFirestore.instance;
+  User? _user;
+  late Future<DocumentSnapshot> _userDocFuture;
+
+  // Controllers for the editable fields
+  final TextEditingController _firstNameController = TextEditingController();
+  final TextEditingController _lastNameController = TextEditingController();
+  final TextEditingController _dobController = TextEditingController();
+
+  bool _controllersInitialized = false;
+
+  @override
+  void initState() {
+    super.initState();
+    _user = _auth.currentUser;
+
+    if (_user != null) {
+      _userDocFuture = _firestore.collection('users').doc(_user!.uid).get();
+    }
+  }
+
+  @override
+  void dispose() {
+    _firstNameController.dispose();
+    _lastNameController.dispose();
+    _dobController.dispose();
+    super.dispose();
+  }
+
+  void _saveProfile() async {
+    if (_user != null) {
+      try {
+        await _firestore.collection('users').doc(_user!.uid).set({
+          'firstName': _firstNameController.text.trim(),
+          'lastName': _lastNameController.text.trim(),
+          'dob': _dobController.text.trim(),
+        }, SetOptions(merge: true));
+        ScaffoldMessenger.of(context).showSnackBar(
+          SnackBar(content: Text('Profile updated')),
+        );
+      } catch (e) {
+        ScaffoldMessenger.of(context).showSnackBar(
+          SnackBar(content: Text('Error updating profile: $e')),
+        );
+      }
+    }
+  }
+
+  @override
+  Widget build(BuildContext context) {
+    return FutureBuilder<DocumentSnapshot>(
+      future: _userDocFuture,
+      builder: (context, snapshot) {
+        if (snapshot.connectionState == ConnectionState.waiting) {
+          // Show a loading indicator while fetching data
+          return Center(child: CircularProgressIndicator());
+        }
+        if (!snapshot.hasData || !snapshot.data!.exists) {
+          // If user data doesn't exist, show a message
+          return Center(child: Text('No user data found'));
+        }
+        Map<String, dynamic> userData =
+            snapshot.data!.data() as Map<String, dynamic>;
+
+        // Initialize controllers if not already initialized
+        if (!_controllersInitialized) {
+          _firstNameController.text = userData['firstName'] ?? '';
+          _lastNameController.text = userData['lastName'] ?? '';
+          _dobController.text = userData['dob'] ?? '';
+          _controllersInitialized = true;
+        }
+
+        return SingleChildScrollView(
+          child: Center(
+            child: Padding(
+              padding: const EdgeInsets.symmetric(horizontal: 24.0),
+              child: Column(
+                // Center all elements
+                mainAxisAlignment: MainAxisAlignment.center,
+                crossAxisAlignment: CrossAxisAlignment.center,
+                children: [
+                  SizedBox(height: 40),
+                  if (userData['photoURL'] != null)
+                    CircleAvatar(
+                      radius: 50,
+                      backgroundImage: NetworkImage(userData['photoURL']),
+                    ),
+                  SizedBox(height: 20),
+                  Text(
+                    userData['displayName'] ?? '',
+                    style: TextStyle(fontSize: 24, fontWeight: FontWeight.bold),
+                    textAlign: TextAlign.center,
+                  ),
+                  Text(
+                    userData['email'] ?? userData['phoneNumber'] ?? '',
+                    style: TextStyle(fontSize: 16),
+                    textAlign: TextAlign.center,
+                  ),
+                  SizedBox(height: 20),
+                  // Editable TextFields for First Name, Last Name, and DOB
+                  TextField(
+                    controller: _firstNameController,
+                    decoration: InputDecoration(labelText: 'First Name'),
+                    textAlign: TextAlign.center,
+                  ),
+                  SizedBox(height: 10),
+                  TextField(
+                    controller: _lastNameController,
+                    decoration: InputDecoration(labelText: 'Last Name'),
+                    textAlign: TextAlign.center,
+                  ),
+                  SizedBox(height: 10),
+                  TextField(
+                    controller: _dobController,
+                    decoration: InputDecoration(labelText: 'Date of Birth'),
+                    textAlign: TextAlign.center,
+                    readOnly: true, // To use a DatePicker
+                    onTap: () async {
+                      // Show DatePicker when tapped
+                      DateTime? pickedDate = await showDatePicker(
+                        context: context,
+                        initialDate: DateTime.tryParse(_dobController.text) ??
+                            DateTime.now(),
+                        firstDate: DateTime(1900),
+                        lastDate: DateTime.now(),
+                      );
+                      if (pickedDate != null) {
+                        setState(() {
+                          _dobController.text =
+                              "${pickedDate.year}-${pickedDate.month.toString().padLeft(2, '0')}-${pickedDate.day.toString().padLeft(2, '0')}";
+                        });
+                      }
+                    },
+                  ),
+                  SizedBox(height: 20),
+                  ElevatedButton(
+                    onPressed: _saveProfile,
+                    child: Text('Save'),
+                  ),
+                  SizedBox(height: 20),
+                  ElevatedButton(
+                    onPressed: () async {
+                      await _auth.signOut();
+                      // Navigate back to the sign-in screen
+                      Navigator.of(context).pushAndRemoveUntil(
+                        MaterialPageRoute(builder: (context) => HomePage()),
+                        (Route<dynamic> route) => false,
+                      );
+                    },
+                    child: Text('Sign Out'),
+                  ),
+                  SizedBox(height: 40),
+                ],
+              ),
+            ),
+          ),
+        );
+      },
+    );
+  }
+}
