+++ pubspec.yaml
@@ -31,6 +31,7 @@ dependencies:
   flutter:
     sdk: flutter
 
+
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
   cupertino_icons: ^1.0.8
@@ -39,6 +40,7 @@ dependencies:
   sign_in_button: ^3.2.0
   google_sign_in: ^6.1.4
   cloud_firestore: ^5.4.5
+  
  
 
 dev_dependencies:
